useAuthListener is central in Header layout, error persists with resetForm() in useAuthListener, meta.json error in main, Logo error on Homepage, forgot password not added, edit/add user photo using URL not added yet
