+++ src/app/dashboard/page.tsx
@@ -1,10 +1,7 @@
 "use client";
 import { useStore } from "@/utils/Store";
-import useAuthListener from "@/utils/useAuthListener";
 
 export default function Dashboard() {
-  useAuthListener();
-
   const userName = useStore((state) => state?.userName);
 
   return (
